same as before

+++ realskywars-api/src/main/java/joserodpt/realskywars/api/achievements/types/RSWAchievementRCoin.java
@@ -75,7 +75,7 @@ public class RSWAchievementRCoin implements RSWAchievement {
 
     @Override
     public ItemStack getItem(RSWPlayer p) {
-        return Itens.createItem(((int) p.getStatistics(this.getType(), false)) >= this.getGoal() ? Material.GREEN_CONCRETE : Material.RED_CONCRETE
+        return Itens.createItem(p.getStatistics(this.getType(), false) >= this.getGoal() ? Material.GREEN_CONCRETE : Material.RED_CONCRETE
                 , 1, TranslatableLine.ACHIEVEMENT_GOAL.get(p).replace("%goal%", String.valueOf(this.getGoal())), Collections.singletonList(TranslatableLine.ACHIEVEMENT_REWARD.get(p).replace("%reward%", this.getReward().toString()) + " " + TranslatableLine.ACHIEVEMENT_NAME_COINS.get(p)));
     }
 }
